% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{connect}
\alias{connect}
\title{Creates a server object that can be passed into other functions for making requests to the PEcAn API}
\usage{
connect(url, username = NULL, password = NULL)
}
\arguments{
\item{url}{Base URL of the PEcAn API Server}

\item{username}{Username}

\item{password}{Password corresponding to the username}
}
\value{
A server object that can be passed into other functions for making requests to the PEcAn API
}
\description{
Creates a server object containing the base URL of the PEcAn API server,
the username & password, which can be simply passed into other functions
to make requests to the PEcAn API.
Currently, there is no validation. In future, we may validate the username-password
pair by hitting the appropriate API endpoint
}
\examples{
server <- connect(url="http://pecan.localhost:80", username="carya", password="illinois")
}
