% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.input.R
\name{download.input}
\alias{download.input}
\title{Download & save the desired input file}
\usage{
download.input(server, input_id, filename = NULL, save_as = NULL)
}
\arguments{
\item{server}{Server object obtained using the connect() function}

\item{input_id}{ID of the PEcAn input to be downloaded}

\item{filename}{Optional filename specified if the id points to a folder instead of file}

\item{save_as}{File name to save the downloaded file as. Default: NULL (same as 'filename' if specified,
else defaults to 'pecan_input_file')}
}
\value{
Response obtained from the \verb{/api/inputs/\{input_id\}} endpoint
}
\description{
Downloads & saves the desired input file.
Hits the \verb{/api/inputs/\{input_id\}} API endpoint
}
\examples{
server <- connect(url="http://localhost:8000", username="carya", password="illinois")

# Download the 'niwot.clim' file (id = 99000000003)
download.input(server, input_id='99000000003', save_as='local.niwot.clim')
}
\author{
Tezan Sahu
}
