% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.run.input.R
\name{download.run.input}
\alias{download.run.input}
\title{Download & save the desired input file for a run}
\usage{
download.run.input(server, run_id, filename, save_as = NULL)
}
\arguments{
\item{server}{Server object obtained using the connect() function}

\item{run_id}{ID of the PEcAn run whose input is needed}

\item{filename}{Name of the input file to be downloaded}

\item{save_as}{File name to save the downloaded file as. Default: NULL (same
name as \code{filename} would be used)}
}
\value{
Response obtained from the \verb{/api/run/\{run_id\}/input/\{filename\}} endpoint
}
\description{
Downloads & saves the desired input file for a run.
Hits the \verb{/api/run/\{run_id\}/input/\{filename\}} API endpoint
}
\examples{
server <- connect(url="http://localhost:8000", username="carya", password="illinois")

# Download the 'sipnet.clim' input for the run with id = '99000000282'
download.run.input(server, run_id=99000000282, filename='sipnet.clim', save_as='test.sipnet.clim')
}
\author{
Tezan Sahu
}
