% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.run.output.R
\name{download.run.output}
\alias{download.run.output}
\title{Download & save the desired output file for a run}
\usage{
download.run.output(server, run_id, filename, save_as = NULL)
}
\arguments{
\item{server}{Server object obtained using the connect() function}

\item{run_id}{ID of the PEcAn run whose output is needed}

\item{filename}{Name of the output file to be downloaded}

\item{save_as}{File name to save the downloaded file as. Default: NULL (same
name as \code{filename} would be used)}
}
\value{
Response obtained from the \verb{/api/run/\{run_id\}/output/\{filename\}} endpoint
}
\description{
Downloads & saves the desired output file for a run.
Hits the \verb{/api/run/\{run_id\}/output/\{filename\}} API endpoint
}
\examples{
server <- connect(url="http://pecan.localhost:80", username="carya", password="illinois")

# Download the 'README.txt' output for the run with id = '99000000282'
download.run.output(server, run_id=99000000282, filename='README.txt', save_as='test.README.txt')
}
\author{
Tezan Sahu
}
