% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.workflow.file.R
\name{download.workflow.file}
\alias{download.workflow.file}
\title{Download & save the desired file for a workflow}
\usage{
download.workflow.file(server, workflow_id, filename, save_as = NULL)
}
\arguments{
\item{server}{Server object obtained using the connect() function}

\item{workflow_id}{ID of the PEcAn workflow whose file is needed}

\item{filename}{Name of the file to be downloaded}

\item{save_as}{File name to save the downloaded file as. Default: NULL (same
name as \code{filename} would be used)}
}
\value{
Response obtained from the \verb{/api/workflows/\{id\}/file/\{filename\}} endpoint
}
\description{
Downloads & saves the desired file for a workflow.
Hits the \verb{/api/workflows/\{id\}/file/\{filename\}} API endpoint
}
\examples{
server <- connect(url="http://localhost:8000", username="carya", password="illinois")

# Download the 'ensemble.ts.99000000017.NPP.2002.2002.Rdata' output file for the workflow with id = 99000000031
download.workflow.file(server, workflow_id=99000000031, filename='ensemble.ts.99000000017.NPP.2002.2002.Rdata')
}
\author{
Tezan Sahu
}
