% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.format.R
\name{get.format}
\alias{get.format}
\title{Get details of PEcAn Model from the database using model id}
\usage{
get.format(server, format_id)
}
\arguments{
\item{server}{Server object obtained using the connect() function}

\item{format_id}{ID of the model to retrieve}
}
\value{
Response obtained from the \verb{/api/formats/\{format_id\}} endpoint
}
\description{
Get details of PEcAn Format from the database using format id.
Hits the \verb{/api/formats/\{format_id\}} API endpoint.
}
\examples{
server <- connect(url="http://pecan.localhost:80", username="carya", password="illinois")

# Get details of the Ameriflux.level4.h format (id = 19)
res <- get.format(server, format_id=19)
}
\author{
Tezan Sahu
}
