% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submit.workflows.csv.R
\name{get.model.id}
\alias{get.model.id}
\title{Get the model ID from the exact name & revision}
\usage{
get.model.id(server, model_name, revision)
}
\arguments{
\item{server}{Server object obtained using the connect() function}

\item{model_name}{Name of the Model}

\item{revision}{Revision of the Model}
}
\value{
ID of the model with exact name & revision passed as parameters
}
\description{
Get the model ID from the exact name & revision
}
\author{
Tezan Sahu
}
