% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.model.R
\name{get.model}
\alias{get.model}
\title{Get details of PEcAn Model from the database using model id}
\usage{
get.model(server, model_id)
}
\arguments{
\item{server}{Server object obtained using the connect() function}

\item{model_id}{ID of the model to retrieve}
}
\value{
Response obtained from the \verb{/api/models/\{model_id\}} endpoint
}
\description{
Get details of PEcAn Model from the database using model id.
Hits the \verb{/api/models/\{model_id\}} API endpoint.
}
\examples{
\dontrun{
server <- connect(url="http://localhost:8000", username="carya", password="illinois")

# Get details of the SIPNET ssr model (id = 1000000022)
res <- get.model(server, model_id=1000000022)
}
}
\author{
Tezan Sahu
}
