% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.models.R
\name{get.models}
\alias{get.models}
\title{Get information about PEcAn Models from the database}
\usage{
get.models(server, model_name = NULL, revision = NULL)
}
\arguments{
\item{server}{Server object obtained using the connect() function}

\item{model_name}{Name of the specific model being searched for. Default: NULL (Returns all models)}

\item{revision}{Revision corresponding to the model being searched for. Default: NULL (Returns all revisions for a model)}
}
\value{
Response obtained from the \verb{/api/models/} endpoint with relevant query parameters
}
\description{
Retrive information about PEcAn Models from the database. Filters
can also be applied based on model_name & revision to get details
about specific models. Hits the \verb{/api/models/} API endpoint with
relevant query parameters
}
\examples{
server <- connect(url="http://localhost:8000", username="carya", password="illinois")

# Get details of all models
res1 <- get.models(server)

# Get details of all 'SIPNET' models
res2 <- get.models(server, "SIPNET")

# Get details of the 'ssr' revision of the 'SIPNET' model
res3 <- get.models(server, "SIPNET", "ssr")
}
