% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.pft.R
\name{get.pft}
\alias{get.pft}
\title{Get details of PEcAn PFT from the database using PFT id}
\usage{
get.pft(server, pft_id)
}
\arguments{
\item{server}{Server object obtained using the connect() function}

\item{pft_id}{ID of the model to retrieve}
}
\value{
Response obtained from the \verb{/api/pfts/\{pft_id\}} endpoint
}
\description{
Get details of PEcAn PFT from the database using PFT id.
Hits the \verb{/api/pfts/\{pft_id\}} API endpoint.
}
\examples{
server <- connect(url="http://pecan.localhost:80", username="carya", password="illinois")

# Get details of the temperate.deciduous PFT (id = 41)
res <- get.pft(server, pft_id=41)
}
\author{
Tezan Sahu
}
