% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.run.R
\name{get.run}
\alias{get.run}
\title{Get the details of a particular PEcAn run using the run ID}
\usage{
get.run(server, run_id)
}
\arguments{
\item{server}{Server object obtained using the connect() function}

\item{run_id}{ID of the PEcAn run whose details are needed}
}
\value{
Response obtained from the \verb{/api/run/\{id\}} endpoint
}
\description{
Obtain the details of a details of a specified PEcAn run.
Hits the \verb{/api/run/\{id\}} API endpoint
}
\examples{
server <- connect(url="https://pecan-tezan.ncsa.illinois.edu", username="carya", password="illinois")

# Get details of run with id = '1002042201'
res <- get.run(server, 1002042201)
}
\author{
Tezan Sahu
}
