% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.runs.R
\name{get.runs}
\alias{get.runs}
\title{Get list of PEcAn Runs that belong to a specific workflow}
\usage{
get.runs(server, workflow_id)
}
\arguments{
\item{server}{Server object obtained using the connect() function}

\item{workflow_id}{ID of the PEcAn workflow for which all runs are to be retrieved}
}
\value{
Response obtained from the \verb{/api/runs/} endpoint with relevant query parameters
}
\description{
Retrive a list of all PEcAn runs from the database that belong to a
specific workflow
}
\examples{
server <- connect(url="http://localhost:8000", username="carya", password="illinois")

# Get details of all runs for the workflow with ID '1000009172'
res <- get.runs(server, "1000009172")
}
