% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.site.R
\name{get.site}
\alias{get.site}
\title{Get details of PEcAn Site from the database using site id}
\usage{
get.site(server, site_id)
}
\arguments{
\item{server}{Server object obtained using the connect() function}

\item{site_id}{ID of the model to retrieve}
}
\value{
Response obtained from the \verb{/api/sites/\{site_id\}} endpoint
}
\description{
Get details of PEcAn Site from the database using site id.
Hits the \verb{/api/sites/\{site_id\}} API endpoint.
}
\examples{
server <- connect(url="http://pecan.localhost:80", username="carya", password="illinois")

# Get details of the WillowCreek(US-WCr) site (id = 676)
res <- get.site(server, site_id=676)
}
\author{
Tezan Sahu
}
