% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.status.R
\name{get.status}
\alias{get.status}
\title{Gets general information about PEcAn & the details of the database host}
\usage{
get.status(server)
}
\arguments{
\item{server}{Server object obtained using the connect() function}
}
\value{
Response obtained from the \verb{/api/status} endpoint
}
\description{
Obtain general information about PEcAn & the details of the database host
Hits the \verb{/api/status} API endpoint
}
\examples{
server <- connect(url="http://localhost:8000", username="carya", password="illinois")
res <- get.status(server)
}
\author{
Tezan Sahu
}
