% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.workflow.status.R
\name{get.workflow.status}
\alias{get.workflow.status}
\title{Get the status of a particular PEcAn workflow using the workflow ID}
\usage{
get.workflow.status(server, workflow_id)
}
\arguments{
\item{server}{Server object obtained using the connect() function}

\item{workflow_id}{ID of the PEcAn workflow whose status are needed}
}
\value{
Response obtained from the \verb{/api/workflow/\{id\}/status} endpoint
}
\description{
Obtain the status of a particular PEcAn workflow by supplying
its ID. Hits the \verb{/api/workflows/\{id\}/status} API endpoint
}
\examples{
server <- connect(url="http://localhost:8000", username="carya", password="illinois")

# Get details of workflow with id = '1000009172'
res <- get.workflow.status(server, workflow_id=1000009172)
}
\author{
Tezan Sahu
}
