% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.workflows.R
\name{get.workflows}
\alias{get.workflows}
\title{Get list of PEcAn Workflows from the database}
\usage{
get.workflows(server, model_id = NULL, site_id = NULL, offset = 0, limit = 50)
}
\arguments{
\item{server}{Server object obtained using the connect() function}

\item{model_id}{ID of the specific model being used in workflows to be retrieved. Default: NULL (Returns workflows using any model)}

\item{site_id}{ID of the specific site being used in workflows to be retrieved. Default: NULL (Returns workflows using any site)}
}
\value{
Response obtained from the \verb{/api/workflows/} endpoint with relevant query parameters
}
\description{
Retrive a list of PEcAn Workflows from the database. Filters
can also be applied based on model_id & site_id to get details
about specific workflows. Hits the \verb{/api/workflows/} API endpoint with
relevant query parameters
}
\examples{
server <- connect(url="http://localhost:8000", username="carya", password="illinois")

# Get list of all workflows
res1 <- get.workflows(server)

# Get list of all workflows using 'SIPNET-ssr' model
res2 <- get.workflows(server, model_id="1000000022")

##' # Get list of all workflows using 'Willow Creek (US-WCr)' as site
res3 <- get.workflows(server, site_id="676")

# Get list of workflow(s) that use 'SIPNET-ssr' model & 'Willow Creek (US-WCr)' as site
res4 <- get.workflows(server, "1000000022", "676")
}
