% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ping.R
\name{ping}
\alias{ping}
\title{Pings the PEcAn API server to see if it is live}
\usage{
ping(server)
}
\arguments{
\item{server}{Server object obtained using the connect() function}
}
\value{
Response obtained from the \verb{/api/ping} endpoint
}
\description{
Pings the PEcAn API server (using the URL passed in the server object)
to see if it is live or not. Hits the \verb{/api/ping} API endpoint
}
\examples{
server <- connect(url="http://pecan.localhost:80", username="carya", password="illinois")
res <- ping(server)
}
\author{
Tezan Sahu
}
