% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_run_vars.R
\name{plot_run_vars}
\alias{plot_run_vars}
\title{Plot the desired variables from the results of a PEcAn run}
\usage{
plot_run_vars(
  server,
  run_id,
  year,
  y_var,
  x_var = "time",
  width = 800,
  height = 600,
  filename = "plot.png"
)
}
\arguments{
\item{server}{Server object obtained using the connect() function}

\item{run_id}{ID of the PEcAn run whose details are needed}

\item{year}{the year this data is for}

\item{width}{the width of the image generated, default is 800 pixels.}

\item{height}{the height of the image generated, default is 600 pixels.}

\item{filename}{File name to save the generated plot, default is 'plot.png'.}

\item{yvar}{the variable to plot along the y-axis.}

\item{xvar}{the variable to plot along the x-axis, by default time is used.}
}
\value{
Response obtained from the \verb{/api/run/\{run_id\}/graph/\{year\}/\{y_var\}} endpoint
}
\description{
Plot the desired variables from the results of a PEcAn run.
Hits the \verb{/api/run/\{run_id\}/graph/\{year\}/\{y_var\}} API endpoint
}
\examples{
server <- connect(url="http://localhost:8000", username="carya", password="illinois")

# Plot the Gross Primary Productivity vs Time for the run with ID '99000000282' for the year 2002
plot_run_vars(server, run_id=99000000282, year=2002, y_var="GPP")

# Plot the Total Respiration vs Soil Respiration for the run with ID '99000000283' for the year 2002 of custom size
plot_run_vars(server, run_id=99000000283, year=2002, y_var="TotalResp", x_var="SoilResp", width=500, height=400)
}
\author{
Tezan Sahu
}
