% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.formats.R
\name{search.formats}
\alias{search.formats}
\title{Search for PEcAn Models from the database}
\usage{
search.formats(server, format_name = "", mimetype = "", ignore_case = TRUE)
}
\arguments{
\item{server}{Server object obtained using the connect() function}

\item{format_name}{Search string for format name. Default: "" (Returns all models)}

\item{ignore_case}{Indicator of case sensitive or case insensitive search}

\item{revision}{Search string for revision. Default: "" (Returns all revisions for a model)}
}
\value{
Response obtained from the \verb{/api/formats/} endpoint with relevant query parameters
}
\description{
Search PEcAn Formats from the database using search string based on
format_name & mimetype Hits the \verb{/api/formats/} API endpoint with
relevant query parameters
}
\examples{
server <- connect(url="http://localhost:8000", username="carya", password="illinois")

# Get details of all formats
res1 <- search.formats(server)

# Get details of all formats containing 'ameriflux' in their name
res2 <- search.formats(server, format_name="ameriflux")

# Get details of the 'ameriflux' formats of 'csv' mimetype
res3 <- search.formats(server, format_name="ameriflux", mimetype="csv")
}
\author{
Tezan Sahu
}
