% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.inputs.R
\name{search.inputs}
\alias{search.inputs}
\title{Search for PEcAn Models from the database}
\usage{
search.inputs(
  server,
  model_id = NULL,
  site_id = NULL,
  format_id = NULL,
  host_id = NULL,
  offset = 0,
  limit = 50
)
}
\arguments{
\item{server}{Server object obtained using the connect() function}

\item{model_id}{ID of the model associated with the input, Default: NULL (Returns all models)}

\item{site_id}{ID of the model associated with the input, Default: NULL (Returns all sites)}

\item{format_id}{ID of the format associated with the input, Default: NULL (Returns all formats)}

\item{host_id}{ID of the host containing the input, Default: NULL (Returns all hosts)}

\item{offset}{The number of workflows to skip before starting to collect the result set.}

\item{limit}{The number of workflows to return (Available values : 10, 20, 50, 100, 500)}
}
\value{
Response obtained from the \verb{/api/inputs/} endpoint with relevant query parameters
}
\description{
Search PEcAn Input files from the database using search based on
model_id & site_id Hits the \verb{/api/inputs/} API endpoint with
relevant query parameters
}
\examples{
\dontrun{
server <- connect(url="http://localhost:8000", username="carya", password="illinois")

# Get details of all inputs
res1 <- search.inputs(server)

# Get details of all inputs for the SIPNET ssr model (id = 1000000022)
res2 <- search.inputs(server, model_id="1000000022")

# Get details of all inputs for the SIPNET ssr model (id = 1000000022) & Niwor Ridge site (id = 772)
res3 <- search.inputs(server, model_id="1000000022", site_id="772")
}
}
\author{
Tezan Sahu
}
