% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.models.R
\name{search.models}
\alias{search.models}
\title{Search for PEcAn Models from the database}
\usage{
search.models(server, model_name = "", revision = "", ignore_case = TRUE)
}
\arguments{
\item{server}{Server object obtained using the connect() function}

\item{model_name}{Search string for model name Default: "" (Returns all models)}

\item{revision}{Search string for revision. Default: "" (Returns all revisions for a model)}

\item{ignore_case}{Indicator of case sensitive or case insensitive search}
}
\value{
Response obtained from the \verb{/api/models/} endpoint with relevant query parameters
}
\description{
Search PEcAn Models from the database using search string based on
model_name & revision. Hits the \verb{/api/models/} API endpoint with
relevant query parameters
}
\examples{
server <- connect(url="https://pecan-tezan.ncsa.illinois.edu", username="carya", password="illinois")

# Get details of all models
res1 <- search.models(server)

# Get details of all models containing 'sip' in their name
res2 <- search.models(server, model_name="sip")

# Get details of the 'ssr' revision of the 'SIPNET' model
res3 <- search.models(server, model_name="SIPNET", revision="ssr")

# Get details of models where name contains "SIP" (case sensitive)
res4 <- search.models(server, model_name="SIPNET", ignore_case=FALSE)
}
\author{
Tezan Sahu
}
