% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.pfts.R
\name{search.pfts}
\alias{search.pfts}
\title{Search for PEcAn PFTs from the database}
\usage{
search.pfts(
  server,
  pft_name = "",
  pft_type = "",
  model_type = "",
  ignore_case = TRUE
)
}
\arguments{
\item{server}{Server object obtained using the connect() function}

\item{pft_name}{Search string for pft name. Default: "" (Returns all pfts)}

\item{pft_type}{PFT type to be filtered. Could be one out of "plant", "cultivar" or "".
Default: "" (Returns PFTs of all types)}

\item{model_type}{Search string for model type associated with PFT. Default: "" (Returns PFTs with any model type)}

\item{ignore_case}{Indicator of case sensitive or case insensitive search}
}
\value{
Response obtained from the \verb{/api/pfts/} endpoint with relevant query parameters
}
\description{
Search PEcAn PFTs from the database using search string based on
PFT name, type & model type. Hits the \verb{/api/pfts/} API endpoint
with relevant query parameters
}
\examples{
server <- connect(url="http://pecan.localhost:80", username="carya", password="illinois")

# Get details of all PFTs
res1 <- search.pfts(server)

# Get details of all PFTs containing 'temperate' in their name
res2 <- search.pfts(server, pft_name="temperate")

# Get details of all PFTs containing of 'cultivar' type
res3 <- search.pfts(server, pft_type="cultivar")

# Get details of PFTs with 'SIPNET' model type
res4 <- search.pfts(server, model_type='SIPNET')
}
\author{
Tezan Sahu
}
