% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.sites.R
\name{search.sites}
\alias{search.sites}
\title{Search for PEcAn Sites from the database}
\usage{
search.sites(server, sitename = "", ignore_case = TRUE)
}
\arguments{
\item{server}{Server object obtained using the connect() function}

\item{sitename}{Search string for site name Default: "" (Returns all sites)}

\item{ignore_case}{Indicator of case sensitive or case insensitive search}
}
\value{
Response obtained from the \verb{/api/sites/} endpoint with relevant query parameters
}
\description{
Search PEcAn Sites from the database using search string based on
sitename. Hits the \verb{/api/sites/} API endpoint with
relevant query parameters
}
\examples{
server <- connect(url="http://localhost:8000", username="carya", password="illinois")

# Get details of all sites
res1 <- search.sites(server)

# Get details of all sites containing 'willow' in their name
res2 <- search.sites(server, sitename="willow")


# Get details of sites where name contains "Willow" (case sensitive)
res4 <- search.sites(server, sitename="Willow", ignore_case=FALSE)
}
