% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submit.workflow.json.R
\name{submit.workflow.json}
\alias{submit.workflow.json}
\title{Submit a JSON file as a PEcAn workflow & obtain the workflow_id}
\usage{
submit.workflow.json(server, jsonFile)
}
\arguments{
\item{server}{Server object obtained using the connect() function}

\item{jsonFile}{JSON file containing the workflow configurations}
}
\value{
Response obtained from the \verb{POST /api/workflow/} endpoint
}
\description{
Submit a JSON file as a PEcAn workflow.
Hits the \verb{POST /api/workflows/} API endpoint
}
\examples{
server <- connect(url="http://pecan.localhost:80", username="carya", password="illinois")
res <- submit.workflow.json(server, "api.sipnet.json")
}
\author{
Tezan Sahu
}
