% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submit.workflow.R
\name{submit.workflow}
\alias{submit.workflow}
\title{Submit a PEcAn workflow using various user-defined parameters
Hits the \verb{POST /api/workflows/} API endpoint.}
\usage{
submit.workflow(
  server,
  model_id,
  site_id,
  pfts,
  start_date,
  end_date,
  inputs,
  meta.analysis = list(),
  ensemble_size = 1,
  sensitivity_variable = "NPP",
  sensitivity.analysis = FALSE,
  notes = NULL,
  workflow_list_mods = list()
)
}
\arguments{
\item{server}{Server object obtained using the connect() function}

\item{model_id}{ID of the model to be used (character)}

\item{site_id}{ID of the site to be used (character)}

\item{pfts}{List of PFTs to be used (list)}

\item{start_date}{Starting date of the analysis (character)}

\item{end_date}{Ending date of the analysis (character)}

\item{inputs}{Inputs to the workflow (including meteorological data, etc.) (object)}

\item{meta.analysis}{Meta-analysis settings object for the workflow. Default: NULL (uses default parameters)}

\item{ensemble_size}{Ensemble size for the workflow. Default: 1}

\item{sensitivity_variable}{Variable for performing sensitivity. Default: NPP}

\item{sensitivity.analysis}{Whether or not to perform a sensitivity analysis. Can also take
a sensitivity setting object as input. Default: FALSE (logical or list)}

\item{notes}{Additional notes that the user need to specify for the submitted workflow. Default: NULL}

\item{workflow_list_mods}{List of additional changes to be applied to the
workflow list. Passed to \code{\link[utils:modifyList]{utils::modifyList()}}. Default = \verb{list() (no changes)}.}
}
\value{
Response obtained from the \verb{POST /api/workflows/} endpoint
}
\description{
Submit a PEcAn workflow using various user-defined parameters
Hits the \verb{POST /api/workflows/} API endpoint.
}
\examples{
server <- connect(url="http://localhost:8000", username="carya", password="illinois")

# Submit a workflow with the SIPNET r136 model (id = 1000000014) for Niwot Ridge site (id = 772) with 
# PFT as 'temperate.coniferous' starting from 01-01-2002 to 31-12-2003 using the input met data with
# id = 99000000003
res <- submit.workflow(server, model_id=1000000014, site_id=772, pfts=c("temperate.coniferous"), start_date="2002-01-01", 
  end_date="2003-12-31", inputs=list(met=list(id=99000000003)))
}
\author{
Tezan Sahu, Alexey Shiklomanov
}
