% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submit.workflow.xml.R
\name{submit.workflow.xml}
\alias{submit.workflow.xml}
\title{Submit an XML file as a PEcAn workflow & obtain the workflow_id}
\usage{
submit.workflow.xml(server, xmlFile)
}
\arguments{
\item{server}{Server object obtained using the connect() function}

\item{xmlFile}{XML file containing the workflow configurations}
}
\value{
Response obtained from the \verb{POST /api/workflow/} endpoint
}
\description{
Submit an XML file as a PEcAn workflow.
Hits the \verb{POST /api/workflows/} API endpoint
}
\examples{
server <- connect(url="http://localhost:8000", username="carya", password="illinois")
res <- submit.workflow.xml(server, "api.sipnet.xml")
}
\author{
Tezan Sahu
}
