% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submit.workflows.csv.R
\name{submit.workflows.csv}
\alias{submit.workflows.csv}
\title{Submit a bunch of PEcAn workflows, the details of which are mentioned in CSV file}
\usage{
submit.workflows.csv(server, csvFile)
}
\arguments{
\item{server}{Server object obtained using the connect() function}

\item{csvFile}{CSV file containing the list of all workflow configurations to be submitted}
}
\value{
Response obtained from the \verb{POST /api/workflow/} endpoint
}
\description{
Submits a bunch of PEcAn workflows, the details of which are mentioned in CSV format (Currently works for SIPNET models)
CSV Template:
Columns: model  revision  met  site_id  pft  start_date  end_date  sensitivity  ensemble  comment
Each row corresponds to one workflow to be submitted
}
\examples{
server <- connect(url="http://pecan.localhost:80", username="carya", password="illinois")
res <- submit.workflows.csv(server, "default_tests.csv")
}
\author{
Tezan Sahu
}
